Cleanup filter resource definition hierarchy.  Fix ModCluster resource description keys (it is registered as a filter, not a handler).
